LayoutTests/http/tests/inspector/network/intercept-request-properties-expected.txt

Test request interception altering request properties.


== Running test suite: Network.interceptWithRequest
-- Running test case: Network.interceptRequest.Method
Triggering load...
Request details:
  URI: /inspector/network/resources/intercept-echo.py
  Response URL: http://127.0.0.1:8000/inspector/network/resources/intercept-echo.py
  Method: POST
  Request Headers:

-- Running test case: Network.interceptRequest.MethodDelete
Triggering load...
Request details:
  URI: /inspector/network/resources/intercept-echo.py
  Response URL: http://127.0.0.1:8000/inspector/network/resources/intercept-echo.py
  Method: DELETE
  Request Headers:

-- Running test case: Network.interceptRequest.MethodNonStandard
Triggering load...
Request details:
  URI: /inspector/network/resources/intercept-echo.py
  Response URL: http://127.0.0.1:8000/inspector/network/resources/intercept-echo.py
  Method: NONSTANDARD
  Request Headers:

-- Running test case: Network.interceptRequest.MethodEmpty
Triggering load...
Request details:
  URI: /inspector/network/resources/intercept-echo.py
  Response URL: http://127.0.0.1:8000/inspector/network/resources/intercept-echo.py
  Method: GET
  Request Headers:

-- Running test case: Network.interceptRequest.URL
Triggering load...
Request details:
  URI: /inspector/network/resources/intercept-echo.py?newURL=value
  Response URL: http://127.0.0.1:8000/inspector/network/resources/intercept-echo.py?newURL=value
  Method: GET
  Request Headers:
  Get:
    newURL: value

-- Running test case: Network.interceptRequest.URL.Regex
Triggering load...
Request details:
  URI: /inspector/network/resources/intercept-echo.py?newURL=value
  Response URL: http://127.0.0.1:8000/inspector/network/resources/intercept-echo.py?newURL=value
  Method: GET
  Request Headers:
  Get:
    newURL: value

-- Running test case: Network.interceptRequest.URLFragment
Triggering load...
Request details:
  URI: /inspector/network/resources/intercept-echo.py
  Response URL: http://127.0.0.1:8000/inspector/network/resources/intercept-echo.py
  Method: GET
  Request Headers:

-- Running test case: Network.interceptRequest.URLEmpty
Triggering load...
Request details:
  URI: /inspector/network/resources/intercept-echo.py
  Response URL: http://127.0.0.1:8000/inspector/network/resources/intercept-echo.py
  Method: GET
  Request Headers:

-- Running test case: Network.interceptRequest.Headers
Triggering load...
Request details:
  URI: /inspector/network/resources/intercept-echo.py
  Response URL: http://127.0.0.1:8000/inspector/network/resources/intercept-echo.py
  Method: GET
  Request Headers:
    X-Value: overridden

-- Running test case: Network.interceptRequest.Headers.Passthrough
Triggering load...
Request details:
  URI: /inspector/network/resources/intercept-echo.py
  Response URL: http://127.0.0.1:8000/inspector/network/resources/intercept-echo.py
  Method: POST
  Request Headers:
    Content-Type: application/x-www-form-urlencoded
  Post:
    value: overridden

-- Running test case: Network.interceptRequest.PostData
Triggering load...
Request details:
  URI: /inspector/network/resources/intercept-echo.py
  Response URL: http://127.0.0.1:8000/inspector/network/resources/intercept-echo.py
  Method: POST
  Request Headers:
    Content-Type: application/x-www-form-urlencoded
  Post:
    value: overridden
